[Doc, BugFix] Fix tensordictmodule tutorial (#819)
